apps/api/pyproject.toml
[project]
name = "trainerlab-api"
version = "0.0.1"
description = "TrainerLab FastAPI backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
    "redis>=5.2.0",
    "python-jose[cryptography]>=3.3.0",
    "typer>=0.15.0",
    "google-auth[requests]>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "slowapi>=0.1.9",
    "google-cloud-tasks>=2.0.0",
    "anthropic>=0.42.0",
    "google-cloud-storage>=2.14.0",
    "openpyxl>=3.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "ty>=0.0.1a7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ASYNC", "N", "S", "C4", "PT"]
ignore = ["S101"]  # Allow assert in tests

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S105", "S106"]  # Allow hardcoded test tokens/secrets

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ty.environment]
python-version = "3.11"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
log_cli = true
log_cli_level = "INFO"

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]
